@@ -3,7 +3,7 @@ import 'package:suche_app/util/constants.dart';
 import 'package:suche_app/views/loginPage/login_page.dart';
 
 class HomePage extends StatefulWidget {
-  HomePage({Key key}) : super(key: key);
+  HomePage({Key? key}) : super(key: key);
 
   @override
   _HomePageState createState() => _HomePageState();
