@@ -4,26 +4,33 @@ import 'package:brasil_fields/brasil_fields.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:mask_text_input_formatter/mask_text_input_formatter.dart';
 import 'package:suche_app/provider/user_provider.dart';
 import 'package:suche_app/util/constants.dart';
 import 'package:intl/intl.dart';
+import 'package:validators/validators.dart';
 
 class RegisterPage extends StatefulWidget {
-  const RegisterPage({Key key}) : super(key: key);
+  const RegisterPage({Key? key}) : super(key: key);
 
   @override
   _RegisterPageState createState() => _RegisterPageState();
 }
 
 class _RegisterPageState extends State<RegisterPage> {
-  String _dropdownValueSexo = null;
-  DateTime _dateTime;
+  String? _dropdownValueSexo;
+  DateTime? _dateTime;
 
   TextEditingController _emailController = TextEditingController();
   TextEditingController _nomeController = TextEditingController();
   TextEditingController _sobrenomeController = TextEditingController();
   TextEditingController _senhaController = TextEditingController();
   TextEditingController _confirmarSenhaController = TextEditingController();
+
+  var _emailMaskFormatter = new MaskTextInputFormatter(mask: '####@#.#', filter: { "#": RegExp(r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+") });
+
+
+  final _formKey = GlobalKey<FormState>();
 
   @override
   Widget build(BuildContext context) {
@@ -60,334 +67,351 @@ class _RegisterPageState extends State<RegisterPage> {
                     horizontal: 40.0,
                     vertical: 80.0,
                   ),
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: <Widget>[
-                      Text(
-                        'Registre - se',
-                        style: TextStyle(
-                          color: Colors.white,
-                          fontFamily: 'OpenSans',
-                          fontSize: 30.0,
-                          fontWeight: FontWeight.bold,
+                  child: Form(
+                    key: _formKey,
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: <Widget>[
+                        Text(
+                          'Registre - se',
+                          style: TextStyle(
+                            color: Colors.white,
+                            fontFamily: 'OpenSans',
+                            fontSize: 30.0,
+                            fontWeight: FontWeight.bold,
+                          ),
                         ),
-                      ),
-                      SizedBox(height: 20.0),
-                      Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: <Widget>[
-                          Text(
-                            'E-mail',
-                            style: kLabelStyle,
-                          ),
-                          SizedBox(height: 10.0),
-                          Container(
-                            alignment: Alignment.centerLeft,
-                            decoration: kBoxDecorationStyle,
-                            height: 60.0,
-                            child: TextField(
-                              controller: _emailController,
-                              keyboardType: TextInputType.emailAddress,
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontFamily: 'OpenSans',
-                              ),
-                              decoration: InputDecoration(
-                                border: InputBorder.none,
-                                contentPadding: EdgeInsets.only(top: 14.0),
-                                prefixIcon: Icon(
-                                  Icons.email,
-                                  color: Colors.white,
-                                ),
-                                hintText: 'Insira o E-mail',
-                                hintStyle: kHintTextStyle,
-                              ),
+                        SizedBox(height: 20.0),
+                        Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: <Widget>[
+                            Text(
+                              'E-mail',
+                              style: kLabelStyle,
                             ),
-                          ),
-                        ],
-                      ),
-                      SizedBox(height: 10.0),
-                      Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: <Widget>[
-                          Text(
-                            'Nome',
-                            style: kLabelStyle,
-                          ),
-                          SizedBox(height: 10.0),
-                          Container(
-                            alignment: Alignment.centerLeft,
-                            decoration: kBoxDecorationStyle,
-                            height: 60.0,
-                            child: TextField(
-                              controller: _nomeController,
-                              keyboardType: TextInputType.name,
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontFamily: 'OpenSans',
-                              ),
-                              decoration: InputDecoration(
-                                border: InputBorder.none,
-                                contentPadding: EdgeInsets.only(top: 14.0),
-                                prefixIcon: Icon(
-                                  Icons.drive_file_rename_outline,
+                            SizedBox(height: 10.0),
+                            Container(
+                              alignment: Alignment.centerLeft,
+                              decoration: kBoxDecorationStyle,
+                              height: 60.0,
+                              child: TextFormField(
+                                controller: _emailController,
+                                validator: (value) {
+                                  if (value == null || value.isEmpty) {
+                                    return 'O e-mail é um campo obrigatório';
+                                  } else if (!isEmail(value)) {
+                                    return 'Insira um e-mail válido';
+                                  }
+                                  return null;
+                                },
+                                keyboardType: TextInputType.emailAddress,
+                                style: TextStyle(
                                   color: Colors.white,
+                                  fontFamily: 'OpenSans',
                                 ),
-                                hintText: 'Insira o Nome',
-                                hintStyle: kHintTextStyle,
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                      SizedBox(height: 10.0),
-                      Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: <Widget>[
-                          Text(
-                            'Sobrenome',
-                            style: kLabelStyle,
-                          ),
-                          SizedBox(height: 10.0),
-                          Container(
-                            alignment: Alignment.centerLeft,
-                            decoration: kBoxDecorationStyle,
-                            height: 60.0,
-                            child: TextField(
-                              controller: _sobrenomeController,
-                              keyboardType: TextInputType.name,
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontFamily: 'OpenSans',
-                              ),
-                              decoration: InputDecoration(
-                                border: InputBorder.none,
-                                contentPadding: EdgeInsets.only(top: 14.0),
-                                prefixIcon: Icon(
-                                  Icons.drive_file_rename_outline,
-                                  color: Colors.white,
-                                ),
-                                hintText: 'Insira o Sobrenome',
-                                hintStyle: kHintTextStyle,
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                      SizedBox(height: 10.0),
-                      Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: <Widget>[
-                          Text(
-                            'Telefone',
-                            style: kLabelStyle,
-                          ),
-                          SizedBox(height: 10.0),
-                          Container(
-                            alignment: Alignment.centerLeft,
-                            decoration: kBoxDecorationStyle,
-                            height: 60.0,
-                            child: TextField(
-                              controller: _confirmarSenhaController,
-                              keyboardType: TextInputType.phone,
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontFamily: 'OpenSans',
-                              ),
-                              decoration: InputDecoration(
-                                border: InputBorder.none,
-                                contentPadding: EdgeInsets.only(top: 14.0),
-                                prefixIcon: Icon(
-                                  Icons.ac_unit,
-                                  color: Colors.white,
-                                ),
-                                hintText: 'Insira seu número',
-                                hintStyle: kHintTextStyle,
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                      SizedBox(height: 10.0),
-                      Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: <Widget>[
-                          Text(
-                            'Gênero',
-                            style: kLabelStyle,
-                          ),
-                          SizedBox(height: 10.0),
-                          Container(
-                            alignment: Alignment.center,
-                            decoration: kBoxDecorationStyle,
-                            height: 60.0,
-                            padding: EdgeInsets.symmetric(horizontal: 12),
-                            child: Row(
-                              children: [
-                                Icon(
-                                  Icons.ac_unit,
-                                  color: Colors.white,
-                                ),
-                                const SizedBox(
-                                  width: 16,
-                                ),
-                                Expanded(
-                                  child: DropdownButton<String>(
-                                    value: _dropdownValueSexo,
-                                    icon: const Icon(Icons.arrow_downward, color: Colors.white),
-                                    iconSize: 24,
-                                    isExpanded: true,
-                                    elevation: 16,
-                                    style: const TextStyle(color: Colors.white),
-                                    dropdownColor: Color(0xFF2B387D),
-                                    underline: Container(
-                                      height: 0,
-                                      color: Colors.white54,
-                                    ),
-                                    hint: _dropdownValueSexo != null
-                                        ? null
-                                        : Text('Selecione o sexo',
-                                        style: kHintTextStyle, textScaleFactor: 1.2),
-                                    onChanged: (String newValue) {
-                                      setState(() {
-                                        _dropdownValueSexo = newValue;
-                                      });
-                                    },
-                                    items: <String>[
-                                      'Masculino',
-                                      'Feminino',
-                                      'Outros',
-                                      'Prefiro não falar'
-                                    ].map<DropdownMenuItem<String>>((String value) {
-                                      return DropdownMenuItem<String>(
-                                        value: value,
-                                        child: Text(
-                                          value,
-                                          textScaleFactor: 1.2,
-                                        ),
-                                      );
-                                    }).toList(),
+                                decoration: InputDecoration(
+                                  border: InputBorder.none,
+                                  contentPadding: EdgeInsets.only(top: 14.0),
+                                  prefixIcon: Icon(
+                                    Icons.email,
+                                    color: Colors.white,
                                   ),
-                                )
-                              ],
-                            ),
-                          ),
-                        ],
-                      ),
-                      SizedBox(height: 10.0),
-                      Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: <Widget>[
-                          Text(
-                            'Data de Nascimento',
-                            style: kLabelStyle,
-                          ),
-                          SizedBox(height: 10.0),
-                          Container(
-                            decoration: kBoxDecorationStyle,
-                            height: 60.0,
-                            child: TextField(
-                              readOnly: true,
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontFamily: 'OpenSans',
-                              ),
-                              decoration: InputDecoration(
-                                border: InputBorder.none,
-                                contentPadding: EdgeInsets.only(top: 14.0),
-                                prefixIcon: Icon(
-                                  Icons.calendar_today,
-                                  color: Colors.white,
+                                  hintText: 'Insira o E-mail',
+                                  hintStyle: kHintTextStyle,
                                 ),
-                                hintText: _dateTime == null
-                                    ? 'Insira a data de nascimento'
-                                    : UtilData.obterDataDDMMAAAA(_dateTime),
-                                hintStyle: kHintTextStyle,
                               ),
-                              onTap: () {
-                                showDatePicker(
-                                  context: context,
-                                  initialDate: DateTime.now(),
-                                  firstDate: DateTime(1920),
-                                  lastDate: DateTime.now(),
-                                ).then((date) {
-                                  setState(() {
-                                    _dateTime = date;
+                            ),
+                          ],
+                        ),
+                        SizedBox(height: 10.0),
+                        Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: <Widget>[
+                            Text(
+                              'Nome',
+                              style: kLabelStyle,
+                            ),
+                            SizedBox(height: 10.0),
+                            Container(
+                              alignment: Alignment.centerLeft,
+                              decoration: kBoxDecorationStyle,
+                              height: 60.0,
+                              child: TextField(
+                                controller: _nomeController,
+                                keyboardType: TextInputType.name,
+                                style: TextStyle(
+                                  color: Colors.white,
+                                  fontFamily: 'OpenSans',
+                                ),
+                                decoration: InputDecoration(
+                                  border: InputBorder.none,
+                                  contentPadding: EdgeInsets.only(top: 14.0),
+                                  prefixIcon: Icon(
+                                    Icons.drive_file_rename_outline,
+                                    color: Colors.white,
+                                  ),
+                                  hintText: 'Insira o Nome',
+                                  hintStyle: kHintTextStyle,
+                                ),
+                              ),
+                            ),
+                          ],
+                        ),
+                        SizedBox(height: 10.0),
+                        Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: <Widget>[
+                            Text(
+                              'Sobrenome',
+                              style: kLabelStyle,
+                            ),
+                            SizedBox(height: 10.0),
+                            Container(
+                              alignment: Alignment.centerLeft,
+                              decoration: kBoxDecorationStyle,
+                              height: 60.0,
+                              child: TextField(
+                                controller: _sobrenomeController,
+                                keyboardType: TextInputType.name,
+                                style: TextStyle(
+                                  color: Colors.white,
+                                  fontFamily: 'OpenSans',
+                                ),
+                                decoration: InputDecoration(
+                                  border: InputBorder.none,
+                                  contentPadding: EdgeInsets.only(top: 14.0),
+                                  prefixIcon: Icon(
+                                    Icons.drive_file_rename_outline,
+                                    color: Colors.white,
+                                  ),
+                                  hintText: 'Insira o Sobrenome',
+                                  hintStyle: kHintTextStyle,
+                                ),
+                              ),
+                            ),
+                          ],
+                        ),
+                        SizedBox(height: 10.0),
+                        Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: <Widget>[
+                            Text(
+                              'Telefone',
+                              style: kLabelStyle,
+                            ),
+                            SizedBox(height: 10.0),
+                            Container(
+                              alignment: Alignment.centerLeft,
+                              decoration: kBoxDecorationStyle,
+                              height: 60.0,
+                              child: TextField(
+                                controller: _confirmarSenhaController,
+                                keyboardType: TextInputType.phone,
+                                style: TextStyle(
+                                  color: Colors.white,
+                                  fontFamily: 'OpenSans',
+                                ),
+                                decoration: InputDecoration(
+                                  border: InputBorder.none,
+                                  contentPadding: EdgeInsets.only(top: 14.0),
+                                  prefixIcon: Icon(
+                                    Icons.ac_unit,
+                                    color: Colors.white,
+                                  ),
+                                  hintText: 'Insira seu número',
+                                  hintStyle: kHintTextStyle,
+                                ),
+                              ),
+                            ),
+                          ],
+                        ),
+                        SizedBox(height: 10.0),
+                        Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: <Widget>[
+                            Text(
+                              'Gênero',
+                              style: kLabelStyle,
+                            ),
+                            SizedBox(height: 10.0),
+                            Container(
+                              alignment: Alignment.center,
+                              decoration: kBoxDecorationStyle,
+                              height: 60.0,
+                              padding: EdgeInsets.symmetric(horizontal: 12),
+                              child: Row(
+                                children: [
+                                  Icon(
+                                    Icons.ac_unit,
+                                    color: Colors.white,
+                                  ),
+                                  const SizedBox(
+                                    width: 16,
+                                  ),
+                                  Expanded(
+                                    child: DropdownButton<String>(
+                                      value: _dropdownValueSexo,
+                                      icon: const Icon(Icons.arrow_downward, color: Colors.white),
+                                      iconSize: 24,
+                                      isExpanded: true,
+                                      elevation: 16,
+                                      style: const TextStyle(color: Colors.white),
+                                      dropdownColor: Color(0xFF2B387D),
+                                      underline: Container(
+                                        height: 0,
+                                        color: Colors.white54,
+                                      ),
+                                      hint: _dropdownValueSexo != null
+                                          ? null
+                                          : Text('Selecione o sexo',
+                                          style: kHintTextStyle, textScaleFactor: 1.2),
+                                      onChanged: (String? newValue) {
+                                        setState(() {
+                                          _dropdownValueSexo = newValue;
+                                        });
+                                      },
+                                      items: <String>[
+                                        'Masculino',
+                                        'Feminino',
+                                        'Outros',
+                                        'Prefiro não falar'
+                                      ].map<DropdownMenuItem<String>>((String value) {
+                                        return DropdownMenuItem<String>(
+                                          value: value,
+                                          child: Text(
+                                            value,
+                                            textScaleFactor: 1.2,
+                                          ),
+                                        );
+                                      }).toList(),
+                                    ),
+                                  )
+                                ],
+                              ),
+                            ),
+                          ],
+                        ),
+                        SizedBox(height: 10.0),
+                        Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: <Widget>[
+                            Text(
+                              'Data de Nascimento',
+                              style: kLabelStyle,
+                            ),
+                            SizedBox(height: 10.0),
+                            Container(
+                              decoration: kBoxDecorationStyle,
+                              height: 60.0,
+                              child: TextField(
+                                readOnly: true,
+                                style: TextStyle(
+                                  color: Colors.white,
+                                  fontFamily: 'OpenSans',
+                                ),
+                                decoration: InputDecoration(
+                                  border: InputBorder.none,
+                                  contentPadding: EdgeInsets.only(top: 14.0),
+                                  prefixIcon: Icon(
+                                    Icons.calendar_today,
+                                    color: Colors.white,
+                                  ),
+                                  hintText: _dateTime == null
+                                      ? 'Insira a data de nascimento'
+                                      : UtilData.obterDataDDMMAAAA(_dateTime!),
+                                  hintStyle: kHintTextStyle,
+                                ),
+                                onTap: () {
+                                  showDatePicker(
+                                    context: context,
+                                    initialDate: DateTime.now(),
+                                    firstDate: DateTime(1920),
+                                    lastDate: DateTime.now(),
+                                  ).then((date) {
+                                    setState(() {
+                                      _dateTime = date;
+                                    });
                                   });
-                                });
-                              },
+                                },
+                              ),
                             ),
-                          ),
-                        ],
-                      ),
-                      SizedBox(height: 10.0),
-                      Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: <Widget>[
-                          Text(
-                            'Senha',
-                            style: kLabelStyle,
-                          ),
-                          SizedBox(height: 10.0),
-                          Container(
-                            alignment: Alignment.centerLeft,
-                            decoration: kBoxDecorationStyle,
-                            height: 60.0,
-                            child: TextField(
-                              controller: _senhaController,
-                              keyboardType: TextInputType.visiblePassword,
-                              obscureText: true,
+                          ],
+                        ),
+                        SizedBox(height: 10.0),
+                        Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: <Widget>[
+                            Text(
+                              'Senha',
+                              style: kLabelStyle,
+                            ),
+                            SizedBox(height: 10.0),
+                            Container(
+                              alignment: Alignment.centerLeft,
+                              decoration: kBoxDecorationStyle,
+                              height: 60.0,
+                              child: TextField(
+                                controller: _senhaController,
+                                keyboardType: TextInputType.visiblePassword,
+                                obscureText: true,
+                                style: TextStyle(
+                                  color: Colors.white,
+                                  fontFamily: 'OpenSans',
+                                ),
+                                decoration: InputDecoration(
+                                  border: InputBorder.none,
+                                  contentPadding: EdgeInsets.only(top: 14.0),
+                                  prefixIcon: Icon(
+                                    Icons.ac_unit,
+                                    color: Colors.white,
+                                  ),
+                                  hintText: 'Insira a Senha',
+                                  hintStyle: kHintTextStyle,
+                                ),
+                              ),
+                            ),
+                          ],
+                        ),
+                        Container(
+                          padding: EdgeInsets.symmetric(vertical: 25.0),
+                          width: double.infinity,
+                          child: ElevatedButton(
+                            style: ElevatedButton.styleFrom(
+                              primary: Colors.white,
+                              padding: EdgeInsets.all(15),
+                              elevation: 5.0,
+                              shape: RoundedRectangleBorder(
+                                borderRadius: BorderRadius.circular(30.0),
+                              ),
+                            ),
+                            onPressed: () async {
+                              // if (_formKey.currentState!.validate()) {
+                                // If the form is valid, display a snackbar. In the real world,
+                                // you'd often call a server or save the information in a database.
+                                ScaffoldMessenger.of(context)
+                                    .showSnackBar(SnackBar(content: Text('Processing Data')));
+                              // }
+                              // final UserProvider _apiClient = new UserProvider();
+                              //
+                              // await _apiClient.createUser(nome: _nomeController.text, sobrenome: _sobrenomeController.text, email: _emailController.text, dataNascimento: _dateTime.toString(), genero: _dropdownValueSexo, password: _senhaController.text, telefone: _confirmarSenhaController.text);
+                              //
+                              // Navigator.pushNamedAndRemoveUntil(context, homeRoute, (route) => false);
+
+                            },
+                            child: Text(
+                              'CADASTRAR',
                               style: TextStyle(
-                                color: Colors.white,
+                                color: Color(0xFF527DAA),
+                                letterSpacing: 1.5,
+                                fontSize: 18.0,
+                                fontWeight: FontWeight.bold,
                                 fontFamily: 'OpenSans',
                               ),
-                              decoration: InputDecoration(
-                                border: InputBorder.none,
-                                contentPadding: EdgeInsets.only(top: 14.0),
-                                prefixIcon: Icon(
-                                  Icons.ac_unit,
-                                  color: Colors.white,
-                                ),
-                                hintText: 'Insira a Senha',
-                                hintStyle: kHintTextStyle,
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                      Container(
-                        padding: EdgeInsets.symmetric(vertical: 25.0),
-                        width: double.infinity,
-                        child: ElevatedButton(
-                          style: ElevatedButton.styleFrom(
-                            primary: Colors.white,
-                            padding: EdgeInsets.all(15),
-                            elevation: 5.0,
-                            shape: RoundedRectangleBorder(
-                              borderRadius: BorderRadius.circular(30.0),
-                            ),
-                          ),
-                          onPressed: () async {
-                            final UserProvider _apiClient = new UserProvider();
-
-                            await _apiClient.createUser(nome: _nomeController.text, sobrenome: _sobrenomeController.text, email: _emailController.text, dataNascimento: _dateTime.toString(), genero: _dropdownValueSexo, password: _senhaController.text, telefone: _confirmarSenhaController.text);
-
-                            Navigator.pushNamedAndRemoveUntil(context, homeRoute, (route) => false);
-
-                          },
-                          child: Text(
-                            'CADASTRAR',
-                            style: TextStyle(
-                              color: Color(0xFF527DAA),
-                              letterSpacing: 1.5,
-                              fontSize: 18.0,
-                              fontWeight: FontWeight.bold,
-                              fontFamily: 'OpenSans',
                             ),
                           ),
                         ),
-                      ),
-                    ],
+                      ],
+                    ),
                   ),
                 ),
               )
