@@ -11,11 +11,11 @@ class HttpAdapter implements HttpClient {
 
   HttpAdapter(this.client);
 
-  Future<dynamic> request({String url, String method, File image, Map body, Map headers, var queryParams}) async {
+  Future<dynamic> request({String? url, String? method, File? image, Map? body, Map? headers, var queryParams}) async {
     final defaultHeaders = headers?.cast<String, String>() ?? {}..addAll({'content-type': 'application/json', 'accept': 'application/json'});
     final jsonBody = body != null ? jsonEncode(body) : null;
     var response = Response('', 500);
-    Uri uri = new Uri.http(AppConfiguration.apiBaseUrl, url, queryParams);
+    Uri uri = new Uri.http(AppConfiguration.apiBaseUrl, url!, queryParams);
     Future<Response> futureResponse = Future<Response>.value(response);
     try {
       if (method == 'post') {
