@@ -2,11 +2,11 @@ import 'dart:io';
 
 abstract class HttpClient {
   Future<dynamic> request({
-    String url,
-    String method,
-    File image,
-    Map body,
-    Map headers,
+    String? url,
+    String? method,
+    File? image,
+    Map? body,
+    Map? headers,
     var queryParams
   });
 }
